Move file associeated collector
